WORKING -- refactor more video logic
