+++ client/src/App.js
@@ -25,16 +25,6 @@ import ChatPanel from './components/ChatPanel';
 import SoundPanel from './components/SoundPanel';
 
 function App({ userName, token, chatRoom }) {
-  const [showVideos, setShowVideos] = useState(true);
-
-  const toggleShowVideos = () => {
-    if (showVideos) {
-      setShowVideos(false);
-      return;
-    }
-    setShowVideos(true);
-  };
-
   const roomName = chatRoom.name;
 
   // const [data, setData] = useState([]); // TODO -- old code, remove
@@ -56,77 +46,6 @@ function App({ userName, token, chatRoom }) {
   //     </div>
   //   );
   // }
-
-  // If we receive an event indicating a track was disabled, execute the code inside
-  function handleTrackDisabled(track) {
-    track.on('disabled', () => {
-      // TODO - render Big Head avatar
-      console.log('Track disabled:');
-      console.log(track);
-    });
-  }
-  // If we receive an event indicating a track was enabled, execute the code inside
-  function handleTrackEnabled(track) {
-    track.on('enabled', () => {
-      // TODO - render Big Head avatar
-      console.log('Track enabled:');
-      console.log(track);
-    });
-  }
-
-  chatRoom.participants.forEach((participant) => {
-    // Display any new media tracks that are subscribed by participants in the room
-    participant.on('trackSubscribed', (track) => {
-      const remoteMediaContainer = document.getElementById('remote-media-div');
-      if (remoteMediaContainer) {
-        remoteMediaContainer.replaceChild(
-          track.attach(),
-          remoteMediaContainer.firstChild
-        );
-      }
-    });
-
-    participant.tracks.forEach((publication) => {
-      // When a new media track is subscribed, attach the listeners to it
-      publication.on('subscribed', handleTrackDisabled);
-      publication.on('subscribed', handleTrackEnabled);
-
-      publication.on('unsubscribed', () => {
-        // TODO - render Big Heads avatar
-        console.log('Publication unsubscribed:');
-        console.log(publication);
-      });
-
-      publication.on('subscribed', () => {
-        // TODO - render Big Heads avatar
-        console.log('Publication subscribed:');
-        console.log(publication);
-      });
-    });
-  });
-
-  useEffect(() => {
-    chatRoom.participants.forEach((participant) => {
-      participant.tracks.forEach((publication) => {
-        // Display the media tracks of participants that are already in the room
-        if (publication.track && document.getElementById('remote-media-div')) {
-          const remoteMediaContainer =
-            document.getElementById('remote-media-div');
-          if (remoteMediaContainer) {
-            remoteMediaContainer.replaceChild(
-              publication.track.attach(),
-              remoteMediaContainer.firstChild
-            );
-          }
-        }
-        // Attach the listeners to every subscribed media track
-        if (publication.isSubscribed) {
-          handleTrackEnabled(publication.track);
-          handleTrackDisabled(publication.track);
-        }
-      });
-    });
-  }, [chatRoom.participants, showVideos]);
 
   // When a new participant connects, display their published media tracks
   chatRoom.on('participantConnected', (participant) => {
@@ -222,8 +141,6 @@ function App({ userName, token, chatRoom }) {
             key={2}
             chatRoom={chatRoom}
             onSelect={() => selectPanel(2)}
-            showVideos={showVideos}
-            toggleShowVideos={toggleShowVideos}
           />
         );
       else if (panel.id === 3)
